listen for external torch state change
used a torch callback to listen for changes in the external state of the torch (such as from other apps, or the notification torch widget). the switches react to the change accordingly.

the torch callback fires up as soon as it's registered so it also knows the previous state of the torches before the app was opened.